- Add vital sign section in AncPatientDetails. - Remove AncPatientDetailsActivity and add isPregnancy check in ViewPagerAdapter.kt and utilize it in PatientDetailsActivity.kt
