+++ android/anc/src/main/AndroidManifest.xml
@@ -33,10 +33,6 @@
             android:theme="@style/AppTheme" />
         <activity android:name="org.smartregister.fhircore.anc.ui.details.PatientDetailsActivity" />
         <activity
-            android:name=".ui.details.bmicompute.BmiQuestionnaireActivity"
-            android:theme="@style/AppTheme" />
-        <activity android:name=".ui.anccare.details.AncDetailsActivity" />
-        <activity
             android:name="org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity"
             android:theme="@style/AppTheme" />
         <activity android:name=".ui.anccare.encounters.EncounterListActivity" />
