+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/family/details/FamilyDetailsActivity.kt
@@ -23,7 +23,6 @@ import org.hl7.fhir.r4.model.Encounter
 import org.smartregister.fhircore.anc.AncApplication
 import org.smartregister.fhircore.anc.data.family.FamilyDetailRepository
 import org.smartregister.fhircore.anc.data.family.model.FamilyMemberItem
-import org.smartregister.fhircore.anc.ui.anccare.details.AncDetailsActivity
 import org.smartregister.fhircore.anc.ui.details.PatientDetailsActivity
 import org.smartregister.fhircore.anc.util.startFamilyMemberRegistration
 import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
@@ -58,19 +57,11 @@ class FamilyDetailsActivity : BaseMultiLanguageActivity() {
   }
 
   private fun onFamilyMemberItemClicked(item: FamilyMemberItem) {
-    if (item.pregnant) {
-      startActivity(
-        Intent(this, AncDetailsActivity::class.java).apply {
-          putExtra(QUESTIONNAIRE_ARG_PATIENT_KEY, item.id)
-        }
-      )
-    } else {
-      startActivity(
-        Intent(this, PatientDetailsActivity::class.java).apply {
-          putExtra(QUESTIONNAIRE_ARG_PATIENT_KEY, item.id)
-        }
-      )
-    }
+    startActivity(
+      Intent(this, PatientDetailsActivity::class.java).apply {
+        putExtra(QUESTIONNAIRE_ARG_PATIENT_KEY, item.id)
+      }
+    )
   }
 
   private fun onAddNewMemberButtonClicked() {
