+++ android/anc/src/main/res/layout/fragment_anc_details.xml
@@ -9,42 +9,12 @@
         android:layout_height="match_parent"
         android:background="@color/backgroundGray">
 
-        <LinearLayout
-            android:id="@+id/linearLayout_details"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@color/colorPrimary"
-            android:orientation="vertical"
-            android:paddingLeft="26dp"
-            android:paddingTop="8dp"
-            android:paddingBottom="16dp"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <TextView
-                android:id="@+id/txtView_patientDetails"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Alberta Tuft, F, 24"
-                android:textColor="@color/white"
-                android:textSize="22sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/txtView_patientId"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="12dp"
-                android:text="Lavington ID: 637846873"
-                android:textColor="@color/light_gray"
-                android:textSize="20sp" />
-
-        </LinearLayout>
 
         <androidx.core.widget.NestedScrollView
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/linearLayout_details"
+            app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_bias="1.0"
             tools:layout_editor_absoluteX="129dp">
