+++ android/anc/src/main/res/menu/profile_menu.xml
@@ -23,6 +23,17 @@
         android:id="@+id/view_past_encounters"
         android:orderInCategory="300"
         android:title="@string/view_past_encounters" />
+
+    <item
+        android:id="@+id/add_vitals"
+        android:orderInCategory="300"
+        android:title="@string/add_vitals" />
+
+    <item
+        android:id="@+id/add_conditions"
+        android:orderInCategory="300"
+        android:title="@string/add_conditions" />
+
     <item
         android:id="@+id/bmi_widget"
         android:orderInCategory="300"
