+++ android/anc/src/test/java/org/smartregister/fhircore/anc/ui/anccare/details/AncDetailsFragmentTest.kt
@@ -54,6 +54,7 @@ import org.smartregister.fhircore.anc.data.model.UpcomingServiceItem
 import org.smartregister.fhircore.anc.data.patient.PatientRepository
 import org.smartregister.fhircore.anc.robolectric.FragmentRobolectricTest
 import org.smartregister.fhircore.anc.shadow.AncApplicationShadow
+import org.smartregister.fhircore.anc.ui.details.PatientDetailsActivity
 import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceDataSource
 
 @ExperimentalCoroutinesApi
@@ -62,7 +63,7 @@ internal class AncDetailsFragmentTest : FragmentRobolectricTest() {
 
   private lateinit var fhirEngine: FhirEngine
   private lateinit var patientDetailsViewModel: AncDetailsViewModel
-  private lateinit var patientDetailsActivity: AncDetailsActivity
+  private lateinit var patientDetailsActivity: PatientDetailsActivity
   private lateinit var patientRepository: PatientRepository
   private lateinit var fragmentScenario: FragmentScenario<AncDetailsFragment>
   private lateinit var patientDetailsFragment: AncDetailsFragment
@@ -101,7 +102,7 @@ internal class AncDetailsFragmentTest : FragmentRobolectricTest() {
       )
 
     patientDetailsActivity =
-      Robolectric.buildActivity(AncDetailsActivity::class.java).create().get()
+      Robolectric.buildActivity(PatientDetailsActivity::class.java).create().get()
     fragmentScenario =
       launchFragmentInContainer(
         factory =
@@ -239,25 +240,6 @@ internal class AncDetailsFragmentTest : FragmentRobolectricTest() {
 
   override fun getFragment(): Fragment {
     return patientDetailsFragment
-  }
-
-  @Test
-  fun testThatDemographicViewsAreUpdated() {
-
-    val item =
-      PatientDetailItem(
-        PatientItem(patientIdentifier = "1", name = "demo", gender = "M", age = "20"),
-        PatientItem(demographics = "2")
-      )
-
-    ReflectionHelpers.callInstanceMethod<Any>(
-      patientDetailsFragment,
-      "handlePatientDemographics",
-      ReflectionHelpers.ClassParameter(PatientDetailItem::class.java, item)
-    )
-
-    Assert.assertEquals("demo, M, 20", patientDetailsFragment.binding.txtViewPatientDetails.text)
-    Assert.assertEquals("2 ID: 1", patientDetailsFragment.binding.txtViewPatientId.text)
   }
 
   @Test
